requirements.txt
asgiref==3.5.2
autopep8==2.0.0
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
